codejam-virtual-keyboard/feat: implement keydown function, update html structure
