+++ src/js/keyclick.js
@@ -1,0 +1,7 @@
+import keysLayout from './keyslayout';
+import removeActiveKey from './removeactivekey';
+
+function keyclickListener(e) {
+
+}
+export default keyclickListener;
