+++ src/js/keyslayout.js
@@ -1,0 +1,76 @@
+const keysLayout = [
+  [
+    ['Backquote', 'ё', 'Ё', '`', '~'],
+    ['Digit1', '1', '!', '1', '!'],
+    ['Digit2', '2', '"', '2', '@'],
+    ['Digit3', '3', '№', '3', '#'],
+    ['Digit4', '4', ';', '4', '$'],
+    ['Digit5', '5', '%', '5', '%'],
+    ['Digit6', '6', ':', '6', '^'],
+    ['Digit7', '7', '?', '7', '&'],
+    ['Digit8', '8', '*', '8', ''],
+    ['Digit9', '9', '(', '9', '('],
+    ['Digit0', '0', ')', '0', ')'],
+    ['Digit-', '-', '-', '-', ''],
+    ['Equal=', '=', '+', '=', '+'],
+    ['Backspace', 'Backspace', 'Backspace', 'Backspace', 'Backspace'],
+  ],
+  [
+    ['Tab', 'Tab', 'Tab', 'Tab', 'Tab'],
+    ['KeyQ', 'й', 'Й', 'q', 'Q'],
+    ['KeyW', 'ц', 'Ц', 'w', 'W'],
+    ['KeyE', 'у', 'У', 'e', 'E'],
+    ['KeyR', 'к', 'К', 'r', 'R'],
+    ['KeyT', 'е', 'Е', 't', 'T'],
+    ['KeyY', 'н', 'Н', 'y', 'Y'],
+    ['KeyU', 'г', 'Г', 'u', 'U'],
+    ['KeyI', 'ш', 'Ш', 'i', 'I'],
+    ['KeyO', 'щ', 'Щ', 'o', 'O'],
+    ['KeyP', 'з', 'З', 'p', 'P'],
+    ['BracketLeft', 'х', 'Х', '[', '{'],
+    ['BracketRight', 'ъ', 'Ъ', ']', '}'],
+    ['Backslash', '\\', '/', '\\', '|'],
+  ],
+  [
+    ['CapsLock', 'CapsLock', 'CapsLock', 'CapsLock', 'CapsLock'],
+    ['KeyA', 'ф', 'Ф', 'a', 'A'],
+    ['KeyS', 'ы', 'Ы', 's', 'S'],
+    ['KeyD', 'в', 'В', 'd', 'D'],
+    ['KeyF', 'а', 'А', 'f', 'F'],
+    ['KeyG', 'п', 'П', 'g', 'G'],
+    ['KeyH', 'р', 'Р', 'h', 'H'],
+    ['KeyJ', 'о', 'О', 'j', 'J'],
+    ['KeyK', 'л', 'Л', 'l', 'L'],
+    ['Semicolon', 'ж', 'Ж', ';', ':'],
+    ['Quote', 'э', 'Э', '\'', '"'],
+    ['Enter', 'Enter', 'Enter', 'Enter', 'Enter'],
+  ],
+
+  [
+    ['ShiftLeft', 'Shift', 'Shift', 'Shift', 'Shift'],
+    ['KeyZ', 'я', 'Я', 'z', 'Z'],
+    ['KeyX', 'ч', 'Ч', 'x', 'X'],
+    ['KeyC', 'с', 'С', 'c', 'C'],
+    ['KeyV', 'м', 'М', 'v', 'V'],
+    ['KeyB', 'и', 'И', 'b', 'B'],
+    ['KeyN', 'т', 'Т', 'b', 'B'],
+    ['KeyM', 'ь', 'Ь', 'm', 'M'],
+    ['Comma', 'б', 'Б', ',', '<'],
+    ['Period', 'ю', 'Ю', '.', '>'],
+    ['Slash', '.', ',', '/', '?'],
+    ['ArrowUp', 'Up', 'Up', 'Up', 'Up'],
+    ['ShiftRight', 'Shift', 'Shift', 'Shift', 'Shift'],
+  ],
+  [
+    ['ControlLeft', 'Ctrl', 'Ctrl', 'Ctrl', 'Ctrl'],
+    ['OSLeft', 'Win', 'Win', 'Win', 'Win'],
+    ['AltLeft', 'Alt', 'Alt', 'Alt', 'Alt'],
+    ['Space', 'Space', 'Space', 'Space', 'Space'],
+    ['AltRight', 'Alt', 'Alt', 'Alt', 'Alt'],
+    ['ControlRight', 'Ctrl', 'Ctrl', 'Ctrl', 'Ctrl'],
+    ['ArrowLeft', 'Left', 'Left', 'Left', 'Left'],
+    ['ArrowDown', 'Down', 'Down', 'Down', 'Down'],
+    ['ArrowRight', 'Right', 'Right', 'Right', 'Right'],
+  ],
+];
+export default keysLayout;
