+++ src/sass/_base.scss
@@ -7,5 +7,5 @@
 }
 
 $fontFuturaBookC: 'FuturaBookC', sans-serif;
-$fontSize: 18px;
+$fontSize: 14px;
 $fontColor: #444;
